Refactor: move embedder and rename vector_search
